+++ test/fixtures/parser/bigquery/declare_variable_with_default.sql
@@ -1,0 +1,3 @@
+declare var5 date default CURRENT_DATE();
+declare var4 int64 default 1;
+declare var3 string default (SELECT "x");